Split out the getEnvInfo() tool helper.
