+++ src/client/pythonEnvironments/info/tool.ts
@@ -1,42 +1,16 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { interpreterInfo as getInterpreterInfoCommand, InterpreterInfoJson } from '../../common/process/internal/scripts';
 import { Architecture } from '../../common/utils/platform';
 import { PythonExecutableInfo, PythonVersion } from '../base/info';
-import { parseVersion } from '../base/info/pythonVersion';
-import { copyPythonExecInfo, PythonExecInfo } from '../exec';
+import { getEnvInfo } from '../base/info/tool';
+import { PythonExecInfo } from '../exec';
 
 export type InterpreterInformation = {
     arch: Architecture;
     executable: PythonExecutableInfo;
     version: PythonVersion;
 };
-
-/**
- * Compose full interpreter information based on the given data.
- *
- * The data format corresponds to the output of the `interpreterInfo.py` script.
- *
- * @param python - the path to the Python executable
- * @param raw - the information returned by the `interpreterInfo.py` script
- */
-function extractInterpreterInfo(python: string, raw: InterpreterInfoJson): InterpreterInformation {
-    const rawVersion = `${raw.versionInfo.slice(0, 3).join('.')}-${raw.versionInfo[3]}`;
-    return {
-        arch: raw.is64Bit ? Architecture.x64 : Architecture.x86,
-        executable: {
-            filename: python,
-            sysPrefix: raw.sysPrefix,
-            mtime: -1,
-            ctime: -1,
-        },
-        version: {
-            ...parseVersion(rawVersion),
-            sysVersion: raw.sysVersion,
-        },
-    };
-}
 
 type ShellExecResult = {
     stdout: string;
@@ -62,29 +36,5 @@ export async function getInterpreterInfo(
     shellExec: ShellExecFunc,
     logger?: Logger,
 ): Promise<InterpreterInformation | undefined> {
-    const [args, parse] = getInterpreterInfoCommand();
-    const info = copyPythonExecInfo(python, args);
-    const argv = [info.command, ...info.args];
-
-    // Concat these together to make a set of quoted strings
-    const quoted = argv.reduce((p, c) => (p ? `${p} "${c}"` : `"${c.replace('\\', '\\\\')}"`), '');
-
-    // Try shell execing the command, followed by the arguments. This will make node kill the process if it
-    // takes too long.
-    // Sometimes the python path isn't valid, timeout if that's the case.
-    // See these two bugs:
-    // https://github.com/microsoft/vscode-python/issues/7569
-    // https://github.com/microsoft/vscode-python/issues/7760
-    const result = await shellExec(quoted, 15000);
-    if (result.stderr) {
-        if (logger) {
-            logger.error(`Failed to parse interpreter information for ${argv} stderr: ${result.stderr}`);
-        }
-        return undefined;
-    }
-    const json = parse(result.stdout);
-    if (logger) {
-        logger.info(`Found interpreter for ${argv}`);
-    }
-    return extractInterpreterInfo(python.pythonExecutable, json);
+    return getEnvInfo(python, shellExec, logger);
 }
